requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.3.0
distlib==0.4.0
distro==1.9.0
fastapi==0.118.0
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
jiter==0.10.0
mypy==1.17.1
mypy_extensions==1.1.0
nodeenv==1.9.1
openai==1.99.9
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.3.0
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.12.9
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.4.0
uvicorn==0.37.0
virtualenv==20.34.0
